backend/requirements.txt
oauthlib
Flask
Flask-JWT-Extended
Flask-OAuthlib
Flask-Migrate
Flask-SQLAlchemy
Flask-Limiter
celery
requests
requests-oauthlib
spacy
cachetools
scikit-learn
transformers
PyPDF2
python-dotenv
docx
langchain
ollama
